chaos_kitten/brain/attack_planner.py
import glob
import logging
import os
from dataclasses import dataclass, field
from typing import Any, Dict, List, Optional

import yaml

logger = logging.getLogger(__name__)


@dataclass
class AttackProfile:
    """Represents a loaded attack profile from a YAML file."""
    name: str
    category: str
    severity: str
    description: str
    payloads: List[str]
    target_fields: List[str]
    success_indicators: Dict[str, Any]
    remediation: str = ""
    references: List[str] = field(default_factory=list)


class AttackPlanner:
    """Plan attacks based on API structure and context.
    
    Uses LLM reasoning to:
    - Understand endpoint semantics
    - Select appropriate attack profiles
    - Plan multi-step attack chains
    - Adapt based on responses
    """
    
    def __init__(self, endpoints: List[Dict[str, Any]], toys_path: str = "toys/") -> None:
        """Initialize the attack planner.
        
        Args:
            endpoints: List of parsed API endpoints
            toys_path: Path to the attack profiles directory
        """
        self.endpoints = endpoints
        self.toys_path = toys_path
        self.attack_profiles: List[AttackProfile] = []
        
        # Configure logging if not already configured
        if not logger.handlers:
            logging.basicConfig(level=logging.INFO)
    
    def load_attack_profiles(self) -> None:
        """Load all attack profiles from the toys directory."""
        # TODO: Load YAML files from toys/
        # raise NotImplementedError("Attack profile loading not yet implemented")
        pass
    
    def plan_attacks(self, endpoint: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Plan attacks for a specific endpoint.
        
        Args:
            endpoint: Endpoint definition from OpenAPI parser.
                      Expected structure:
                      {
                          "path": "/api/login",
                          "method": "post",
                          "parameters": [...],
                          "requestBody": {...} 
                      }
            
        Returns:
            List of planned attacks with payloads and expected behaviors:
            [
                {
                    "profile_name": str,
                    "endpoint": str,
                    "method": str,
                    "field": str,
                    "location": str (query/path/body),
                    "payloads": list[str],
                    "expected_indicators": dict,
                    "severity": str
                },
                ...
            ]
        """
        # MVP: Simple rule-based stub
        attacks = []
        path = endpoint.get("path", "")
        method = endpoint.get("method", "GET")
        
        # Simple heuristic: If it takes parameters, try SQL injection
        params = endpoint.get("parameters", [])
        body = endpoint.get("requestBody", {})
        
        if params or body:
            attacks.append({
                "type": "sql_injection",
                "name": "Basic SQLi Probe",
                "description": "Injects a basic SQL payload to test for errors",
                "payload": {"q": "' OR 1=1 --"}, # Simplified payload assumption
                "target_param": "q" if params else "body",
                "expected_status": 500
            })
            
        return attacks
    
    def reason_about_field(self, field_name: str, field_type: str) -> str:
        """Use LLM to reason about potential vulnerabilities for a field.
        
        Example:
            field_name="age", field_type="integer"
            Returns: "I'll test negative numbers, zero, extremely large values, and strings"
        
        Args:
            field_name: Name of the field
            field_type: Data type of the field
            
        Returns:
            Reasoning about what to test
        """
        # TODO: Implement LLM reasoning in future iteration
        return f"Standard testing for {field_type} field '{field_name}'"
